requirements_speech.txt
appdirs==1.4.3
gTTS==1.1.8
gTTS-token==1.1.1
packaging==16.8
pip==9.0.1
pocketsphinx==0.1.3
PyAudio==0.2.11
pyparsing==2.2.0
pyttsx==1.1
requests==2.13.0
setuptools==34.3.3
six==1.10.0
SpeechRecognition==3.6.3
wheel==0.29.0
python-vlc
